refactor: 프로필 이미지 컴포넌트 함수분리
Github issue #208
